Restructure notebook cell in components folder
