+++ src/components/index.tsx
@@ -1,13 +1,9 @@
 import BorderContainer from './BorderContainer';
-import CellToolbar from './CellToolbar';
-import CodeCell from './CodeCell';
 import ColoredIconButton from './ColoredIconButton';
 import GoogleSignInButton from './GoogleSignInButton';
 import Header from './Header';
 import IconTextButton from './IconTextButton';
-import MarkdownCell from './MarkdownCell';
 import NotebookCell from './NotebookCell';
-import OutputCell from './OutputCell';
 import PopoverDropdown from './PopoverDropdown';
 import RegularIconButton from './RegularIconButton';
 import StatusIndicator from './StatusIndicator';
@@ -16,15 +12,11 @@ import UserAvatar from './UserAvatar';
 
 export {
   BorderContainer,
-  CellToolbar,
-  CodeCell,
   ColoredIconButton,
   GoogleSignInButton,
   Header,
   IconTextButton,
-  MarkdownCell,
   NotebookCell,
-  OutputCell,
   PopoverDropdown,
   RegularIconButton,
   StatusIndicator,
